added:logic to save in MySql database

+++ employeepayroll/src/main/java/com/bridgelabz/employeepayroll/repository/IEmployeeRepository.java
@@ -3,5 +3,5 @@ package com.bridgelabz.employeepayroll.repository;
 import com.bridgelabz.employeepayroll.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IEmployeeRepository extends JpaRepository<Employee,Long> {
+public interface IEmployeeRepository extends JpaRepository<Employee,Integer> {
 }
